# Conflicts:
#	gradle.properties
#	lib_licences/licenseReport.csv
#	lib_licences/licenseReport.html
#	lib_licences/licenseReport.json

+++ src/main/java/de/geheimagentnr1/world_pre_generator/elements/queues/tasks/PrinterSubTask.java
@@ -21,7 +21,7 @@ public class PrinterSubTask extends TimedSubTask {
 	
 	private LocalDateTime old_time;
 	
-	private int old_chunks;
+	private long old_chunks;
 	
 	public PrinterSubTask( PregenTaskQueue _queue ) {
